+++ ui/ui.go
@@ -2450,7 +2450,7 @@ func Build() (*gowid.App, error) {
 	analysisMenuItems := []menuutil.SimpleMenuItem{
 		menuutil.SimpleMenuItem{
 			Txt: "Capture file properties",
-			Key: gowid.MakeKey('c'),
+			Key: gowid.MakeKey('p'),
 			CB: func(app gowid.IApp, w gowid.IWidget) {
 				analysisMenu.Close(app)
 				startCapinfo(app)
@@ -2462,6 +2462,14 @@ func Build() (*gowid.App, error) {
 			CB: func(app gowid.IApp, w gowid.IWidget) {
 				analysisMenu.Close(app)
 				startStreamReassembly(app)
+			},
+		},
+		menuutil.SimpleMenuItem{
+			Txt: "Conversations",
+			Key: gowid.MakeKey('c'),
+			CB: func(app gowid.IApp, w gowid.IWidget) {
+				analysisMenu.Close(app)
+				openConvsUi(app)
 			},
 		},
 	}
@@ -2951,6 +2959,7 @@ func Build() (*gowid.App, error) {
 	menuPathMax = []interface{}{0, 6}
 
 	buildStreamUi()
+	buildFilterConvsMenu()
 
 	mainView = appkeys.New(mainViewNoKeys, mainKeyPress)
 
@@ -3015,6 +3024,8 @@ func Build() (*gowid.App, error) {
 	app.RegisterMenu(analysisMenu)
 	app.RegisterMenu(generalMenu)
 	app.RegisterMenu(conversationMenu)
+	app.RegisterMenu(filterConvsMenu1)
+	app.RegisterMenu(filterConvsMenu2)
 
 	gowid.SetFocusPath(mainview, mainviewPaths[0], app)
 	gowid.SetFocusPath(altview1, altview1Paths[0], app)